dividing modules.
