+++ src/Web/Direct/PersistedInfo.hs
@@ -1,0 +1,45 @@
+{-# LANGUAGE DeriveGeneric     #-}
+{-# LANGUAGE OverloadedStrings #-}
+
+module Web.Direct.PersistedInfo (
+    PersistedInfo(..)
+  , serializePersistedInfo
+  , deserializePersistedInfo
+  ) where
+
+import           Data.Aeson                       (FromJSON, ToJSON,
+                                                   fieldLabelModifier)
+import qualified Data.Aeson                       as Json
+import qualified Data.ByteString.Lazy             as B
+import qualified Data.Char                        as Char
+import qualified Data.Text                        as T
+import           GHC.Generics                     (Generic)
+
+----------------------------------------------------------------
+
+data PersistedInfo = PersistedInfo {
+    persistedInfoDirectAccessToken :: !T.Text
+  , persistedInfoIdfv              :: !T.Text
+  } deriving (Eq, Show, Generic)
+
+instance FromJSON PersistedInfo where
+  parseJSON = Json.genericParseJSON deriveJsonOptions
+
+instance ToJSON PersistedInfo where
+  toJSON = Json.genericToJSON deriveJsonOptions
+  toEncoding = Json.genericToEncoding deriveJsonOptions
+
+serializePersistedInfo :: PersistedInfo -> B.ByteString
+serializePersistedInfo = Json.encode
+
+deserializePersistedInfo :: B.ByteString -> Either String PersistedInfo
+deserializePersistedInfo = Json.eitherDecode
+
+deriveJsonOptions :: Json.Options
+deriveJsonOptions = Json.defaultOptions
+    { fieldLabelModifier = firstLower . drop (T.length "PersistedInfo")
+    }
+
+firstLower :: String -> String
+firstLower (x : xs) = Char.toLower x : xs
+firstLower _        = error "firstLower: Assertion failed: empty string"
