+++ src/Web/Login.hs
@@ -1,0 +1,1 @@
+module Web.Direct.Login where
